refactor desktop sidebar
